Generate random constants using a simple LCG instead of ChaCha20
This lets us drop the `rand` dependency, making the library more lightweiight, and enabling a WebAssembly target in the future.

If there was concern about the low quality of the random constants, we could switch to a custom ChaCha implementation, which is still fairly easy to implement and test. This should be revisited before production use.